% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listChelsaCmip5.R
\name{list_chelsa_cmip5}
\alias{list_chelsa_cmip5}
\title{Get list of GCM, RCP and climate projections from CHELSA-CMIP5 datasets}
\usage{
list_chelsa_cmip5()
}
\value{
a list of 4 vectors:
  - a vector of RCP names
  - a vector of GCM names
  - a vector of GCM-RCP combination for 2041-2060
  - a vector of GCM-RCP combination for 2061-2080
}
\description{
Return a list of GCM, RCP and climate projections from CHELSA-CMIP5 datasets (\url{http://chelsa-climate.org/}).
}
\examples{

cmip5_infos <- list_chelsa_cmip5()

length(cmip5_infos)
## [1] 4

names(cmip5_infos)
## [1] "rcp" "gcm" "2041-2060" "2061-2080"

cmip5_infos$rcp
## [1] "rcp26" "rcp45" "rcp60" "rcp85"
}
\author{
Nicolas Casajus, \email{nicolas.casajus@fondationbiodiversite.fr}
}
