% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgetChelsa.R
\name{wget_chelsa}
\alias{wget_chelsa}
\title{Bulk download of CHELSA climate datasets}
\usage{
wget_chelsa(path = ".", vars = NULL, horizons = NULL, gcms = NULL,
  rcps = NULL)
}
\arguments{
\item{path}{a character vector of the path to stored dowloaded files.}

\item{vars}{a character/numeric vector of bioclimatic variable(s) id (from 1 to 19).}

\item{horizons}{a character vector of time slices ("1979-2013", "2041-2060", "2061-2080").}

\item{gcms}{a character vector of CMIP5 GCM names (see \url{http://chelsa-climate.org/future/} for a complete list). Only required if \code{time = "2041-2060"} or \code{time = "2061-2080"}.}

\item{rcps}{a character vector of RCP ("rcp26", "rcp45", "rcp60", "rcp85"). Only required if \code{time = "2041-2060"} or \code{time = "2061-2080"}.}
}
\value{
No console output (files are written on the hard drive).
}
\description{
This function downloads climate normals from the Chelsa database (\url{http://chelsa-climate.org/}).
User can select specific bioclimatic variables, time slices, GCMs and RCPs.
}
\note{
For future horizons, if no values RCP are specified, projections for all available RCP will downloaded for the chosen GCMs. If no values GCM are specified, projections for all available GCM will downloaded for the chosen RCPs.
}
\examples{

wget_chelsa(
  vars     = c(1, 7, 12, 15),
  horizons = c("1979-2013", "2041-2060"),
  gcms     = c("CESM1-BGC", "MIROC5"),
  rcps     = "rcp85"
)
}
\author{
Nicolas Casajus, \email{nicolas.casajus@fondationbiodiversite.fr}
}
